ProjetHitema/requirements.txt
selenium
webdriver-manager
pandas
sqlalchemy
pymysql
mysql-connector-python